# backend/doc/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: ProdCom-E API
  description: An e-commerce website to deliver best and authentic gadget items.
  contact:
    name: Md. Sifytul Karim
    email: sifytul09@gmail.com
...
servers:
  - description: localhost
    url: http://localhost:4000
paths:
...
  '/categories':
    get:
      description: This route will retrieve all the categories
      tags:
        - Category
      responses:
        '200':
          description: success response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  categories:
                    type: array
                    items:
                      allOf:
                        - type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            category_name:
                              type: string
                              example: phone
                            description:
                              type: string
                              example: This is a phone category.
                            image_url:
                              type: string
                              example: s3.categoryimage.jpeg
                        - $ref: '#/components/schemas/Timestamp'
                          links:
                            type: object
                            properties:
                              self:
                                type: string
                                example: /categories/1
        '500':
          $ref: '#/components/responses/500'

  # Category - Admin Endpoints
  '/admin/categories':
    post:
      tags:
        - Category
      description: This route will create a category with provided payload.
      security:
        - bearerAuth: []
      requestBody:
        description: category payload
        content:
          'application/json':
            schema:
              type: object
              properties:
                category_name:
                  type: string
                  example: phone
                description:
                  type: string
                  example: This is a phone category.
              required:
                - category_name
      responses:
        '201':
          description: success response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: category successfully created

        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  '/admin/categories/{id}':
    patch:
      tags:
        - Category
      description: This route will update a category based on categoryId
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/categoryIdParam'
      requestBody:
        description: category update payload
        content:
          'application/json':
            schema:
              type: object
              properties:
                category_name:
                  type: string
                  example: phone
                description:
                  type: string
                  example: This is a phone category.
      responses:
        '200':
          $ref: '#/components/responses/200'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Category
      description: This route will delete a category based on categoryId
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/categoryIdParam'
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  # Reviews Endpoints
...
components:
  responses:
...
    '401':
      description: Unauthorized
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Please sign in to perform the action.
    '403':
      description: Forbidden
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: You are not authorized to see these resources
    '500':
      description: Server error
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Internal server Error
...
    '204':
      description: successful deletion of the resource
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: Successfully performed the delete action
    '200':
      description: update success response
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: successfully updated
  parameters:
...
    categoryIdParam:
      in: path
      name: id
      required: true
      description: category ID
      schema:
        type: integer
        format: int32
        example: 2
...
  schemas:
...
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          example: '2023-9-3 03:03:00'
        updatedAt:
          type: string
          example: '2023-9-3 03:03:00'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT